+++ src/js/input.js
@@ -206,16 +206,19 @@ export default {
         return value || '';
     },
     /**
+     * Finds a form control that is not a nested setvalue/xforms-value-changed directive
+     * 
      * @param {string} name
      * @param {number} index
      * @return {Element} found element
      */
-    find( name, index ) {
+    find( name, index = 0 ) {
         let attr = 'name';
         if ( this.form.view.html.querySelector( `input[type="radio"][data-name="${name}"]:not(.ignore)` ) ) {
             attr = 'data-name';
         }
-        const question = this.getWrapNodes( this.form.view.html.querySelectorAll( `[${attr}="${name}"]` ) )[ index ];
+        const selector = `[${attr}="${name}"]:not([data-event="xforms-value-changed"])`;
+        const question = this.getWrapNodes( this.form.view.html.querySelectorAll( selector ) )[ index ];
 
         return question ? question.querySelector( `[${attr}="${name}"]:not(.ignore)` ) : null;
     },