+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 [Unreleased]
 ----------------------
+##### Added
+- Support for multiple setvalue/xforms-value-changed actions under the same form control.
+
 ##### Fixed
 - Border shown when printing a Grid Theme form.